requirements.txt
gTTS
playsound
playsound3
sounddevice
vosk
opencv-python
ultralytics
torch
numpy
mtcnn
keras-facenet
requests
pytesseract
Pillow
easyocr
kagglehub
tensorflow
pandas
matplotlib
scikit-learn